Seed Postgres db with 1-5million
